rewrote checksource and foil read-in in process_foil_data
